Rewards: Split out Point System
Map all point related items together by point system

+++ database/2_tables/point_usage.sql
@@ -17,7 +17,7 @@ SET sql_notes = 1;
 CREATE TABLE point_usage (
     point_usage_id int    NOT NULL AUTO_INCREMENT,
     store_id int    NOT NULL ,
-    yen_per_point double(16,16)    NOT NULL ,
+    yen_per_point decimal(8,6)    NOT NULL ,
     credit_card_id int    NOT NULL ,
     update_time datetime NOT NULL,
     update_user varchar(100) CHARACTER SET utf8 COLLATE utf8_general_ci NOT NULL,
@@ -28,7 +28,7 @@ CREATE TABLE point_usage (
 CREATE TABLE point_usage_history (
     point_usage_id int    NOT NULL ,
     store_id int    NOT NULL ,
-    yen_per_point double(16,16)    NOT NULL ,
+    yen_per_point decimal(8,6)   NOT NULL ,
     credit_card_id int    NOT NULL ,
     time_beg datetime NOT NULL,
     time_end datetime NULL,
